# Improve mempool performance when switching to new peak