--- a/src/consensus/blockchain.py
+++ b/src/consensus/blockchain.py
@@ -118,12 +118,14 @@ class Blockchain(BlockchainInterface):
         """
         Initializes the state of the Blockchain class from the database.
         """
-        height_to_hash, sub_epoch_summaries = await self.block_store.get_sub_block_dicts()
+        height_to_hash, sub_epoch_summaries = await self.block_store.get_peak_sub_height_dicts()
         self.__sub_height_to_hash = height_to_hash
         self.__sub_epoch_summaries = sub_epoch_summaries
         self.__sub_blocks = {}
         self.__sub_heights_in_cache = {}
-        sub_blocks, peak = await self.block_store.get_sub_blocks_from_peak(self.constants.SUB_BLOCKS_CACHE_SIZE)
+        sub_blocks, peak = await self.block_store.get_sub_block_records_close_to_peak(
+            self.constants.SUB_BLOCKS_CACHE_SIZE
+        )
         for sub_block in sub_blocks.values():
             self.add_sub_block(sub_block)
 
@@ -561,7 +563,7 @@ class Blockchain(BlockchainInterface):
         """
         if self.peak_height is None:
             return
-        sub_blocks = await self.block_store.get_sub_block_in_range(
+        sub_blocks = await self.block_store.get_sub_block_records_in_range(
             max(fork_point - self.constants.SUB_BLOCKS_CACHE_SIZE, uint32(0)), fork_point
         )
         for sub_block in sub_blocks.values():
@@ -601,10 +603,10 @@ class Blockchain(BlockchainInterface):
         self.clean_sub_block_record(peak.sub_block_height - self.constants.SUB_BLOCKS_CACHE_SIZE)
 
     async def get_sub_block_records_in_range(self, start: int, stop: int) -> Dict[bytes32, SubBlockRecord]:
-        return await self.block_store.get_sub_block_in_range(start, stop)
+        return await self.block_store.get_sub_block_records_in_range(start, stop)
 
     async def get_header_blocks_in_range(self, start: int, stop: int) -> Dict[bytes32, HeaderBlock]:
-        return await self.block_store.get_headers_in_range(start, stop)
+        return await self.block_store.get_header_blocks_in_range(start, stop)
 
     async def get_sub_block_from_db(self, header_hash: bytes32) -> Optional[SubBlockRecord]:
         if header_hash in self.__sub_blocks: