--- a/src/full_node/block_store.py
+++ b/src/full_node/block_store.py
@@ -143,6 +143,27 @@ class BlockStore:
         await cursor.close()
         return [FullBlock.from_bytes(row[0]) for row in rows]
 
+    async def get_header_blocks_in_range(
+        self,
+        start: int,
+        stop: int,
+    ) -> Dict[bytes32, HeaderBlock]:
+
+        formatted_str = (
+            f"SELECT header_hash,block from full_blocks WHERE sub_height >= {start} and sub_height <= {stop}"
+        )
+
+        cursor = await self.db.execute(formatted_str)
+        rows = await cursor.fetchall()
+        await cursor.close()
+        ret: Dict[bytes32, HeaderBlock] = {}
+        for row in rows:
+            header_hash = bytes.fromhex(row[0])
+            full_block: FullBlock = FullBlock.from_bytes(row[1])
+            ret[header_hash] = full_block.get_block_header()
+
+        return ret
+
     async def get_sub_block_record(self, header_hash: bytes32) -> Optional[SubBlockRecord]:
         cursor = await self.db.execute(
             "SELECT sub_block from sub_block_records WHERE header_hash=?",
@@ -174,28 +195,7 @@ class BlockStore:
                 peak = header_hash
         return ret, peak
 
-    async def get_headers_in_range(
-        self,
-        start: int,
-        stop: int,
-    ) -> Dict[bytes32, HeaderBlock]:
-
-        formatted_str = (
-            f"SELECT header_hash,block from full_blocks WHERE sub_height >= {start} and sub_height <= {stop}"
-        )
-
-        cursor = await self.db.execute(formatted_str)
-        rows = await cursor.fetchall()
-        await cursor.close()
-        ret: Dict[bytes32, HeaderBlock] = {}
-        for row in rows:
-            header_hash = bytes.fromhex(row[0])
-            full_block: FullBlock = FullBlock.from_bytes(row[1])
-            ret[header_hash] = full_block.get_block_header()
-
-        return ret
-
-    async def get_sub_block_in_range(
+    async def get_sub_block_records_in_range(
         self,
         start: int,
         stop: int,
@@ -219,7 +219,9 @@ class BlockStore:
 
         return ret
 
-    async def get_sub_blocks_from_peak(self, blocks_n: int) -> Tuple[Dict[bytes32, SubBlockRecord], Optional[bytes32]]:
+    async def get_sub_block_records_close_to_peak(
+        self, blocks_n: int
+    ) -> Tuple[Dict[bytes32, SubBlockRecord], Optional[bytes32]]:
         """
         Returns a dictionary with all sub_blocks that have sub_height >= peak sub_height - blocks_n, as well as the
         peak header hash.
@@ -241,7 +243,7 @@ class BlockStore:
             ret[header_hash] = SubBlockRecord.from_bytes(row[1])
         return ret, bytes.fromhex(row[0])
 
-    async def get_sub_block_dicts(self) -> Tuple[Dict[uint32, bytes32], Dict[uint32, SubEpochSummary]]:
+    async def get_peak_sub_height_dicts(self) -> Tuple[Dict[uint32, bytes32], Dict[uint32, SubEpochSummary]]:
         """
         Returns a dictionary with all sub blocks, as well as the header hash of the peak,
         if present.