--- a/src/types/mempool_item.py
+++ b/src/types/mempool_item.py
@@ -12,6 +12,7 @@ class MempoolItem:
     fee_per_cost: float
     fee: uint64
     cost_result: CostResult
+    spend_bundle_name: bytes32
 
     def __lt__(self, other):
         # TODO test to see if it's < or >
@@ -19,4 +20,4 @@ class MempoolItem:
 
     @property
     def name(self) -> bytes32:
-        return self.spend_bundle.name()
+        return self.spend_bundle_name