--- a/tests/core/consensus/test_weight_proof.py
+++ b/tests/core/consensus/test_weight_proof.py
@@ -307,7 +307,7 @@ class TestWeightProof:
         connection = await aiosqlite.connect("path to db")
         block_store: BlockStore = await BlockStore.create(connection)
         sub_blocks, peak = await block_store.get_sub_block_records()
-        headers = await block_store.get_headers_in_range(0, 100225)
+        headers = await block_store.get_header_blocks_in_range(0, 100225)
         sub_height_to_hash = {}
         sub_epoch_summaries = {}
         peak = await block_store.get_full_blocks_at([100225])