From my tests using a large TX spending 800 coins, this took the tx validation time on the second time validating from 3 seconds to 70ms. On the current testnet, each block was taking about 6 seconds to add due to refreshing the mempool, this now takes around 800ms.

* Cache signature validations (only validate first time we see the tx)
* Cache spend bundle names (only calculate first time we see the tx)
* Use additions_for_npc for faster computation of additions
* Batch fetch header blocks from disk